python_services/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.3
colorama==0.4.6
cryptography==47.0.0
distro==1.9.0
exceptiongroup==1.3.1
fastapi==0.136.1
filetype==1.2.0
google-auth==2.49.2
google-genai==1.74.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
langchain==1.2.16
langchain-core==1.3.2
langchain-google-genai==4.2.2
langchain-protocol==0.0.14
langgraph==1.1.10
langgraph-checkpoint==4.0.3
langgraph-prebuilt==1.0.13
langgraph-sdk==0.3.13
langsmith==0.7.38
numpy==2.2.6
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.2
pandas==2.3.3
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.13.3
pydantic_core==2.46.3
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.33.1
requests-toolbelt==1.0.0
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
starlette==1.0.0
tenacity==9.1.4
threadpoolctl==3.6.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.46.0
websockets==16.0
xxhash==3.7.0
zstandard==0.25.0
